feat(applesauce): Add an important feature to the applesauce widget
Users want the option to pick the ratio of Granny Smith to Washington Red in their applesauce. 57%
of users are on opposite ends of the apple mix spectrum and would not be satisfied by the same
combination.

BREAKING CHANGE: There is no longer a default applesauce mix. Users must specify a ratio, even if
it's the existing 50/50 mix.

fix #1